Add timestamp linking and remove feed icon
- Add relative time (5m, 2h, 6d) with link to original post in inspector header
- Display webfinger and linked timestamp next to author name in inspector
- Remove time link from feed stage (clicking opens inspector instead)
- Remove icon from Feed menu item in sidebar
- Create shared getRelativeTimeShort utility with i18n support
- Use dateI18n for localized fallback dates older than 6 days
