+++ src/social-web/components/fields/metadata/index.tsx
@@ -2,36 +2,8 @@ import { __ } from '@wordpress/i18n';
 import { decodeEntities } from '@wordpress/html-entities';
 import type { Field } from '@wordpress/dataviews';
 import { useSettings } from '../../../contexts/settings-context';
-import type { FeedPost } from '../../types';
-
-// Helper function to format relative time (max 6 days)
-// dateString should be in GMT/UTC format
-function getRelativeTime( dateString: string ): string {
-	const now = Date.now();
-	// Ensure the date string is parsed as UTC by adding 'Z' if not present
-	const utcDateString = dateString.endsWith( 'Z' ) ? dateString : dateString + 'Z';
-	const date = new Date( utcDateString );
-	const diffMs = now - date.getTime();
-	const diffMinutes = Math.floor( diffMs / ( 1000 * 60 ) );
-	const diffHours = Math.floor( diffMs / ( 1000 * 60 * 60 ) );
-	const diffDays = Math.floor( diffMs / ( 1000 * 60 * 60 * 24 ) );
-
-	// If the date is in the future or just happened, show recent time
-	if ( diffMinutes < 1 ) {
-		return '0m';
-	} else if ( diffMinutes < 60 ) {
-		return `${ diffMinutes }m`;
-	} else if ( diffHours < 24 ) {
-		return `${ diffHours }h`;
-	} else if ( diffDays <= 6 ) {
-		return `${ diffDays }d`;
-	}
-	return date.toLocaleDateString( undefined, {
-		year: 'numeric',
-		month: 'short',
-		day: 'numeric',
-	} );
-}
+import type { FeedPost } from '../../../types';
+import { getRelativeTimeShort } from '../../../utils';
 
 export const metadataField: Field< FeedPost > = {
 	id: 'metadata',
@@ -41,7 +13,7 @@ export const metadataField: Field< FeedPost > = {
 	getValue: ( { item }: { item: FeedPost } ) => {
 		const author = item.actor_info?.name || '';
 		const dateToUse = item.date_gmt || item.date;
-		const relativeTime = dateToUse ? getRelativeTime( dateToUse ) : '';
+		const relativeTime = dateToUse ? getRelativeTimeShort( dateToUse ) : '';
 		return `${ author } · ${ relativeTime }`;
 	},
 	render: ( { item }: { item: FeedPost } ) => {
@@ -50,8 +22,7 @@ export const metadataField: Field< FeedPost > = {
 		const avatarUrl = item.actor_info?.icon || '';
 		// Use date_gmt for accurate time calculations, fallback to date
 		const dateToUse = item.date_gmt || item.date;
-		const relativeTime = dateToUse ? getRelativeTime( dateToUse ) : '';
-		const postLink = item.link || '';
+		const relativeTime = dateToUse ? getRelativeTimeShort( dateToUse ) : '';
 
 		return (
 			<div className="activitypub-feed-post-meta">
@@ -64,18 +35,10 @@ export const metadataField: Field< FeedPost > = {
 					} }
 				/>
 				<span className="author">{ name }</span>
-				{ relativeTime && postLink && (
+				{ relativeTime && (
 					<>
 						<span className="separator">·</span>
-						<a
-							href={ postLink }
-							target="_blank"
-							rel="noopener noreferrer"
-							className="date"
-							onClick={ ( e ) => e.stopPropagation() }
-						>
-							{ relativeTime }
-						</a>
+						<span className="date">{ relativeTime }</span>
 					</>
 				) }
 			</div>
